refactor(layout): extract vertical text detection
Move CJK vertical text extraction and vertical block ordering into a dedicated layout module so the main layout entrypoint handles fewer concerns.
